requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cloudinary==1.25.0
crispy-forms-gds==0.2.2
dj-database-url==0.5.0
Django==3.2.3
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-heroku==0.3.1
django-js-asset==1.2.2
django-model-utils==4.1.1
django-timestampedmodel==0.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.2.0
psycopg2==2.8.6
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
